+++ stroom-core-shared/src/main/java/stroom/query/shared/FetchSuggestionsRequest.java
@@ -16,18 +16,26 @@
 
 package stroom.query.shared;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import stroom.datasource.api.v2.AbstractField;
 import stroom.docref.DocRef;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import javax.validation.constraints.NotNull;
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class FetchSuggestionsRequest {
+
+    @NotNull
     @JsonProperty
     private final DocRef dataSource;
+
+    @NotNull
     @JsonProperty
     private final AbstractField field;
+
     @JsonProperty
     private final String text;